requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
cobs==1.1.3               # via -r requirements.in
idna==2.9                 # via requests
influxdb==5.3.0           # via -r requirements.in
msgpack==0.6.1            # via influxdb
pyserial==3.4             # via -r requirements.in
python-dateutil==2.8.1    # via influxdb
pytz==2020.1              # via influxdb
requests==2.23.0          # via influxdb
six==1.14.0               # via influxdb, python-dateutil
urllib3==1.25.9           # via requests
